requirements.txt
Pillow~=8.3.1
web3~=5.23.0